refactor(individual-tracker): extract top-bar stat functions to top-bar-stats.ts; move tests
- Move accumulatePlayerStats, computeTopBarStats, and all their helpers
  (getTopBarStatLabel, computeSeriesWonLoss, computeKdaValue,
  formatTopBarStatOption) to a new top-bar-stats.ts file for easier
  future migration to shared/
- Move topBarStats tests to tests/top-bar-stats.test.ts; remove
  describe label milestone prefix; use Partial<Stats["CoreStats"]> for
  core-stats override type instead of ReturnType<typeof aFakeCoreStatsWith>
